+++ modules/brj-stylesheet/includes/frontend.js.php
@@ -1,0 +1,11 @@
+(function($){
+
+    var editor = ace.edit("stylesheet-<?php echo $id; ?>-frontend-editor");
+    editor.setTheme("ace/theme/dawn");
+    editor.getSession().setMode({path:"ace/mode/css", inline:true});
+    editor.getSession().setUseWrapMode(true);
+    editor.setOption('displayIndentGuides', true);
+    editor.setShowPrintMargin(false);
+    editor.setReadOnly(true);
+
+})(jQuery);